pyproject.toml
[project]
name = "dog-breed-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "saylock",email = "cntqpb@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "typer (>=0.21.0,<0.22.0)",
    "dvc (>=3.65.0,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "pytorch-lightning (>=2.6.0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "mlflow (>=3.8.1,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]

[tool.poetry.scripts]
dog-breed-bot = "dog_breed_bot.commands:app"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
